requirements.txt
annotated-types==0.6.0
anyio==4.3.0
autopep8==2.0.4
click==8.1.7
dnspython==2.6.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
